requirements.txt
alembic==1.5.4
attrs==20.3.0
autopep8==1.5.5
beautifulsoup4==4.11.1
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
coverage==6.3.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
html2text==2020.1.16
idna==2.10
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.9.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.22.4
packaging==20.9
pandas==1.4.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycurl==7.45.1
pyenchant==3.2.2
pygame==2.1.2
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.12.1
pytest-integration==0.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2022.1
requests==2.25.1
six==1.15.0
soupsieve==2.3.2.post1
SQLAlchemy==1.3.23
toml==0.10.2
tomli==2.0.1
urllib3==1.26.5
Werkzeug==1.0.1
wikipedia==1.4.0
Wikipedia-API==0.5.4
wonderwords==2.2.0
wptools==0.4.17
zipp==3.8.0
